+++ client/app/app.js
@@ -4,6 +4,7 @@ import ngAria from 'angular-aria'
 import uiRouter from 'angular-ui-router';
 import uiBootstrap from 'angular-ui-bootstrap';
 import 'ng-pattern-restrict/src/ng-pattern-restrict';
+import 'ngstorage/ngStorage';
 import Partials from './partials/partials';
 import Pages from './pages/pages';
 import Models from './models/models';
@@ -15,6 +16,7 @@ angular.module('app', [
   uiRouter,
   uiBootstrap,
   'ngPatternRestrict',
+  'ngStorage',
   Partials.name,
   Pages.name,
   Models.name